feat: Move `TritonProgramSpecification` to feature flag `spec`
Problem:
 1. Integration tests (living in the `tests/` directory) cannot import structs,
    traits, and functions guarded by the `test` flag.
 2. Public trait `TritonProgramSpecification` is guarded by the `test` flag.
 :. Therefore, integration tests cannot cover Triton programs that come with a
    specification.

Solution: add a new feature flag `spec` and declare `TritonProgramSpecification`
if either `spec` or `test` is on.

This change anticipates the addition of integration tests that need this trait.
