+++ neptune-core/Cargo.toml
@@ -62,6 +62,9 @@ tokio-console = ["dep:console-subscriber"]
 # the client side.
 mock-rpc = []
 
+# defines and implements TritonProgramSpecification for various programs
+spec = []
+
 [dependencies]
 
 # note: arbitrary, proptest, proptest-arbitrary-interop are duplicated in [dev-dependencies]
@@ -162,6 +165,7 @@ libp2p = { version = "0.56.0", features = [
 whoami = "2.0.0"
 
 [dev-dependencies]
+neptune-cash = { path = "../neptune-core", features = ["arbitrary-impls", "spec"] }
 
 # note: arbitrary, proptest, proptest-arbitrary-interop are duplicated in [dev-dependencies]
 # because they are optional deps but required for unit tests.
