+++ neptune-core/src/protocol/consensus/block/validity.rs
@@ -91,7 +91,7 @@ impl TritonProgram for PrincipalBlockValidationLogic {
 #[cfg_attr(coverage_nightly, coverage(off))]
 mod tests {
     use super::*;
-    use crate::protocol::proof_abstractions::tasm::program::tests::TritonProgramSpecification;
+    use crate::protocol::proof_abstractions::tasm::program::spec::TritonProgramSpecification;
 
     impl TritonProgramSpecification for PrincipalBlockValidationLogic {
         fn source(&self) {
