+++ neptune-core/src/protocol/consensus/block/validity/correct_control_parameter_update.rs
@@ -51,7 +51,7 @@ impl TritonProgram for CorrectControlParameterUpdate {
 #[cfg_attr(coverage_nightly, coverage(off))]
 pub mod tests {
     use super::*;
-    use crate::protocol::proof_abstractions::tasm::program::tests::TritonProgramSpecification;
+    use crate::protocol::proof_abstractions::tasm::program::spec::TritonProgramSpecification;
 
     impl TritonProgramSpecification for CorrectControlParameterUpdate {
         fn source(&self) {
