+++ neptune-core/src/protocol/consensus/block/validity/correct_mmr_update.rs
@@ -51,7 +51,7 @@ impl TritonProgram for CorrectMmrUpdate {
 #[cfg_attr(coverage_nightly, coverage(off))]
 mod tests {
     use super::*;
-    use crate::protocol::proof_abstractions::tasm::program::tests::TritonProgramSpecification;
+    use crate::protocol::proof_abstractions::tasm::program::spec::TritonProgramSpecification;
 
     impl TritonProgramSpecification for CorrectMmrUpdate {
         fn source(&self) {
