+++ neptune-core/src/protocol/consensus/transaction/validity/kernel_to_outputs.rs
@@ -358,24 +358,14 @@ impl TritonProgram for KernelToOutputs {
     }
 }
 
-#[cfg(test)]
-#[cfg_attr(coverage_nightly, coverage(off))]
-mod tests {
-    use proptest::prop_assert_eq;
-    use proptest::strategy::Strategy;
-    use proptest::test_runner::TestRunner;
-    use rand::random;
-    use tasm_lib::triton_vm;
-    use test_strategy::proptest;
-
+#[cfg(any(test, feature = "spec"))]
+mod spec {
     use super::*;
-    use crate::protocol::consensus::transaction::utxo::Utxo;
+
+    use crate::api::export::AdditionRecord;
+    use crate::api::export::Utxo;
     use crate::protocol::proof_abstractions::tasm::builtins as tasm;
-    use crate::protocol::proof_abstractions::tasm::program::tests::test_program_snapshot;
-    use crate::protocol::proof_abstractions::tasm::program::tests::TritonProgramSpecification;
-    use crate::triton_vm::proof::Claim;
-    use crate::triton_vm::stark::Stark;
-    use crate::util_types::mutator_set::addition_record::AdditionRecord;
+    use crate::protocol::proof_abstractions::tasm::program::spec::TritonProgramSpecification;
     use crate::util_types::mutator_set::commit;
 
     impl TritonProgramSpecification for KernelToOutputs {
@@ -423,6 +413,23 @@ mod tests {
             tasm::tasmlib_io_write_to_stdout___digest(salted_output_utxos_hash);
         }
     }
+}
+
+#[cfg(test)]
+#[cfg_attr(coverage_nightly, coverage(off))]
+mod tests {
+    use proptest::prop_assert_eq;
+    use proptest::strategy::Strategy;
+    use proptest::test_runner::TestRunner;
+    use rand::random;
+    use tasm_lib::triton_vm;
+    use test_strategy::proptest;
+
+    use super::*;
+    use crate::protocol::proof_abstractions::tasm::program::spec::TritonProgramSpecification;
+    use crate::protocol::proof_abstractions::tasm::program::tests::test_program_snapshot;
+    use crate::triton_vm::proof::Claim;
+    use crate::triton_vm::stark::Stark;
 
     #[proptest(cases = 30)]
     fn kernel_to_outputs_proptest(
